Session: new HipChat-specific session class.
This supports HipChat-specific session auth, which doesn't require
scraping the web page for info.  This can also automatically fetch an
oauth token, so we don't need to scrape for that either.

The old auth auth scraping code has been moved over to the new session
class as a fallback in case the new flow doesn't work.